Extract first shared action model for the NCP
